# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bidding/nearby:
    get:
      tags: [Bidding]
      operationId: getBiddingNearby
      summary: List active bidding rides within driver's radius.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: radiusKm
          required: false
          schema:
            type: number
            minimum: 0.5
            maximum: 25
            default: 5
      responses:
        '200':
          description: Nearby bidding rides sorted by distance ascending.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BiddingNearbyResponse' }
        '403': { description: Not an approved driver }
        '409': { description: No live driver position available }
  /bidding/offers:
    post:
      tags: [Bidding]
      operationId: createBiddingOffer
      summary: Driver posts (or counter-offers) a bid on a bidding ride.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBiddingOfferBody' }
      responses:
        '201':
          description: Bid created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BiddingOfferResponse' }
        '400': { description: Invalid input or bid violates fare bounds }
        '403': { description: Driver not approved or not online }
        '404': { description: Ride not found }
        '409': { description: Ride is no longer accepting bids }
        '410': { description: Bidding deadline has passed }
  /bidding/offers/{bidId}/withdraw:
    post:
      tags: [Bidding]
      operationId: withdrawBiddingOffer
      summary: Driver withdraws an active bid.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bidId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Bid cancelled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BiddingOfferResponse' }
        '403': { description: Not the bid owner }
        '404': { description: Bid not found }
        '409': { description: Bid is not active }
  /bidding/posts/{rideId}/accept-offer:
    post:
      tags: [Bidding]
      operationId: acceptBiddingOffer
      summary: Rider accepts one driver's offer; other active bids are rejected.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: rideId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcceptBiddingOfferBody' }
      responses:
        '200':
          description: Bid accepted; ride moves to driver_arriving.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptBiddingOfferResponse' }
        '403': { description: Not the ride owner }
        '404': { description: Ride or bid not found }
        '409': { description: Ride is no longer in bidding state }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BiddingNearbyRide:
      type: object
      required:
        - id
        - pickupLabel
        - pickupAddress
        - dropoffLabel
        - dropoffAddress
        - estimatedDistanceKm
        - estimatedDurationMin
        - distanceKm
        - createdAt
      properties:
        id: { type: string, format: uuid }
        pickupLabel: { type: string }
        pickupAddress: { type: string }
        dropoffLabel: { type: string }
        dropoffAddress: { type: string }
        pickupLat: { type: number, nullable: true }
        pickupLng: { type: number, nullable: true }
        initialFare: { type: number, nullable: true }
        vehicleClass: { type: string, nullable: true }
        vehicleTypeId: { type: string, format: uuid, nullable: true }
        estimatedDistanceKm: { type: number }
        estimatedDurationMin: { type: integer }
        distanceKm: { type: number, description: "Pickup distance from the driver's current position, in kilometers." }
        biddingExpiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    BiddingNearbyResponse:
      type: object
      required: [rides]
      properties:
        rides:
          type: array
          items: { $ref: '#/components/schemas/BiddingNearbyRide' }
    BiddingBid:
      type: object
      required: [id, rideId, driverId, amount, etaMin, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        rideId: { type: string, format: uuid }
        driverId: { type: string, format: uuid }
        amount: { type: number }
        etaMin: { type: integer }
        note: { type: string, nullable: true }
        status:
          type: string
          enum: [active, accepted, rejected, cancelled, expired]
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateBiddingOfferBody:
      type: object
      required: [rideId, amount, etaMin]
      properties:
        rideId: { type: string, format: uuid }
        amount: { type: number, exclusiveMinimum: 0, maximum: 100000 }
        etaMin: { type: integer, minimum: 1, maximum: 60 }
        note: { type: string, maxLength: 200 }
    BiddingOfferResponse:
      type: object
      required: [bid]
      properties:
        bid: { $ref: '#/components/schemas/BiddingBid' }
    AcceptBiddingOfferBody:
      type: object
      required: [bidId]
      properties:
        bidId: { type: string, format: uuid }
    AcceptBiddingOfferResponse:
      type: object
      required: [ride, bid]
      properties:
        ride:
          allOf:
            - $ref: '#/components/schemas/AcceptedBiddingRide'
        bid: { $ref: '#/components/schemas/BiddingBid' }
    AcceptedBiddingRide:
      type: object
      description: |
        Ride row after the rider accepted a bid. Status is `driver_arriving`,
        `acceptedBidId` and `acceptedDriverId` point at the winning offer.
        Includes the full `bids` array so the rider client can stay in sync
        without a follow-up GET.
      required:
        - id
        - riderId
        - pickupLabel
        - pickupAddress
        - dropoffLabel
        - dropoffAddress
        - estimatedDistanceKm
        - estimatedDurationMin
        - status
        - acceptedBidId
        - acceptedDriverId
        - bids
        - createdAt
      properties:
        id: { type: string, format: uuid }
        riderId: { type: string, format: uuid }
        pickupLabel: { type: string }
        pickupAddress: { type: string }
        dropoffLabel: { type: string }
        dropoffAddress: { type: string }
        pickupLat: { type: number, nullable: true }
        pickupLng: { type: number, nullable: true }
        dropoffLat: { type: number, nullable: true }
        dropoffLng: { type: number, nullable: true }
        routePolyline: { type: string, nullable: true }
        estimatedDistanceKm: { type: number }
        estimatedDurationMin: { type: integer }
        status:
          type: string
          enum: [bidding, driver_arriving, in_progress, completed, cancelled, queued, assigned_next]
        acceptedBidId: { type: string, format: uuid, nullable: true }
        acceptedDriverId: { type: string, format: uuid, nullable: true }
        finalAmount: { type: number, nullable: true }
        initialFare: { type: number, nullable: true }
        biddingExpiresAt: { type: string, format: date-time, nullable: true }
        bids:
          type: array
          items: { $ref: '#/components/schemas/BiddingBid' }
        createdAt: { type: string, format: date-time }